==> notebook/01_data_preparation.ipynb <==
import pandas as pd
import numpy as np
import networkx as nx
import pickle

# Demographic Profile Data

# boston_pop = pd.read_csv ('Boston/bigquery_BUA_population.csv', index_col = False)
# boston_pop['person_id'] = boston_pop['person_id'].astype(float)
# # only keep the columns we need
# boston_pop = boston_pop[['person_id','age','individual_income','household_size','family_structure','vehicles']]
# boston_pop.head(3)

# Trip Data

# ##Import trip info
# trips_sa = pd.read_csv ('Boston/BUA_trips_bga_sa_amenitiesadded2.csv', index_col = False)
# trips_th = pd.read_csv ('Boston/BUA_trips_bga_th_amenitiesadded2.csv', index_col = False)
# trips_th.head(2)

# cond1 = {
#     'BIKING': 'Biking',
#     'CARPOOL':'Car',
#     'ON_DEMAND_AUTO':'Car',
#     'PRIVATE_AUTO':'Car',
#     'PUBLIC_TRANSIT':'Public_transport',
#     'WALKING':'Walking'
# }

# trips_sa['mode'] = trips_sa['mode'].map(cond1)
# trips_th['mode'] = trips_th['mode'].map(cond1)
# trips_sa=trips_sa[trips_sa['mode'].notnull()]
# trips_th=trips_th[trips_th['mode'].notnull()]
# trips_sa.sort_index(axis = 1)
# trips_th.sort_index(axis = 1)

# # only keep the columns we need
# trips_sa = trips_sa[['activity_id','person_id','travel_purpose','top_amenity','sub_amenity','top_category','sub_category','mode','distance_miles','duration_minutes','location_name']]
# trips_th  = trips_th[['activity_id','person_id','travel_purpose','top_amenity','sub_amenity','top_category','sub_category','mode','distance_miles','duration_minutes','location_name']]
# trips_sa.head(2)

# trips_th_with_pop = trips_th.merge(boston_pop, how = 'inner', on = 'person_id')
# trips_th_with_pop.head(2)

# trips_th_with_pop.to_csv('data/orig/trips_th_with_pop.csv',index = False)

# Create Behavior Graph 

Nodes: 
- Person
    - person_id
    - age_group
    - income_group
    - household_size
    - family_structure
    - vehicles

- Desire
    - travel_purpose
        - Eat
        - Recreation
        - Shop
        - Errands
        - Other
        - School
        - Work

- Intension
    - top_amenity/top_category
    - sub_amenity/sub_category
    - mode
    - distance_miles
    - duration_minutes
    - location


Links:

Person ---want_to---> Desire ---choose_to---> Intension

import pandas as pd
import numpy as np
import networkx as nx
import pickle

trips_th_with_pop = pd.read_csv('../data/orig/trips_th_with_pop.csv',index_col = False)
trips_th_with_pop = trips_th_with_pop[trips_th_with_pop['travel_purpose'] == 'Eat']
# shuffle the data
trips_th_with_pop  = trips_th_with_pop.sample(frac=1).reset_index(drop=True)


# remove the rows where family structure equals to 'GQ'
trips_th_with_pop = trips_th_with_pop[trips_th_with_pop['family_structure']!='GQ']
# remove the rows where vehicles equals to 'GQ'
trips_th_with_pop = trips_th_with_pop[trips_th_with_pop['vehicles']!='GQ']
trips_th_with_pop['vehicles'] = trips_th_with_pop['vehicles'].replace('zero',0)
trips_th_with_pop.head(2)

person_props = ['age','individual_income','household_size','family_structure','vehicles']

desire_prop = ['travel_purpose']

intension_prop = ['top_amenity','sub_amenity','mode','distance_miles','duration_minutes','location_name']

# Prepare train data

person_df = pd.DataFrame(columns = ['id','age','individual_income','household_size','family_structure','vehicles', 'description'])
desire_df = pd.DataFrame(columns = ['id', 'desire', 'description'])
intention_df = pd.DataFrame(columns = ['id','target_amenity','mode','distance_miles','duration_minutes','location_name', 'description'])

want_to_edge_df = pd.DataFrame(columns = ['source','target','type'])
go_to_edge_df = pd.DataFrame(columns = ['source','target','type'])

for idx,row in trips_th_with_pop.iterrows():
    # if idx < 10000:
    #     continue
    person_id = row['person_id']
    person_description = f"A {row['age']} year old person, living in a {row['family_structure']} family with {row['household_size']} members. The person has {row['vehicles']} vehicles and an annual income of {row['individual_income']} dollars."
    
    desire = row['travel_purpose']
    desire_description = f"This person wants to {row['travel_purpose']}"


    intention_id = row['activity_id']
    intention_description = f"This person goes to {row['location_name']} by {row['mode']}, this place is in the amenity category of {row['top_amenity']}/{row['sub_amenity']}. The distance is {row['distance_miles']} miles. It takes {row['duration_minutes']} minutes."


    person_df.loc[len(person_df)] = [f"Person_{person_id}",row['age'],row['individual_income'],row['household_size'],row['family_structure'],row['vehicles'],person_description]

    desire_df.loc[len(desire_df)] = [f"Desire_{person_id}_{desire}",row['travel_purpose'],desire_description]

    intention_df.loc[len(intention_df)] = [f"Internsion_{intention_id}",row['top_amenity']+'/'+row['sub_amenity'],row['mode'],row
    ['distance_miles'],row['duration_minutes'],row['location_name'],intention_description]

    want_to_edge_df.loc[len(want_to_edge_df)] = [f"Person_{person_id}",f"Desire_{person_id}_{desire}",'want_to']

    go_to_edge_df.loc[len(go_to_edge_df)] = [f"Desire_{person_id}_{desire}",f"Internsion_{intention_id}",'go_to']

    if idx>1000:
        break

person_df = person_df.drop_duplicates()
desire_df = desire_df.drop_duplicates()
intention_df = intention_df.drop_duplicates()
want_to_edge_df = want_to_edge_df.drop_duplicates()
go_to_edge_df = go_to_edge_df.drop_duplicates()
print("number of person nodes: ",len(person_df))
print("number of desire nodes: ",len(desire_df))
print("number of intension nodes: ",len(intention_df))
print("number of want_to edges: ",len(want_to_edge_df))
print("number of go_to edges: ",len(go_to_edge_df))

dataset = 'train'
person_df.to_csv(f'data/{dataset}/1K/person.csv',index = False)
desire_df.to_csv(f'data/{dataset}/1K/desire.csv',index = False)
intention_df.to_csv(f'data/{dataset}/1K/intention.csv',index = False)
want_to_edge_df.to_csv(f'data/{dataset}/1K/want_to_edge.csv',index = False)
go_to_edge_df.to_csv(f'data/{dataset}/1K/go_to_edge.csv',index = False)

# Prepare test data

import pandas as pd
import numpy as np
import networkx as nx
import pickle

trips_th_with_pop = pd.read_csv('../data/orig/trips_th_with_pop.csv',index_col = False)
# shuffle the data
trips_th_with_pop  = trips_th_with_pop.sample(frac=1).reset_index(drop=True)
# remove the rows where family structure equals to 'GQ'
trips_th_with_pop = trips_th_with_pop[trips_th_with_pop['family_structure']!='GQ']
# remove the rows where vehicles equals to 'GQ'
trips_th_with_pop = trips_th_with_pop[trips_th_with_pop['vehicles']!='GQ']
trips_th_with_pop['vehicles'] = trips_th_with_pop['vehicles'].replace('zero',0)
trips_th_with_pop.head(2)


income_bins = [-float('inf'), 0, 30000, 60000, 90000, 120000, float('inf')]
income_labels = ['Debt', 'Low', 'Moderate', 'High', 'Very High', 'Ultra High']
age_bins = [-float('inf'), 18, 30, 40, 50, 60, float('inf')]
age_labels = ['Teen', 'Young Adult', 'Adult', 'Middle Age', 'Senior', 'Elderly']

trips_th_with_pop['income_group'] = pd.cut(trips_th_with_pop['individual_income'], bins=income_bins, labels=income_labels)
trips_th_with_pop['age_group'] = pd.cut(trips_th_with_pop['age'], bins=age_bins, labels=age_labels)
trips_th_with_pop.head(2)

duration_step = 5
distance_step = 0.5

trips_th_with_pop['target_amenity'] = trips_th_with_pop['top_amenity']+'/'+trips_th_with_pop['sub_amenity']
trips_th_with_pop['distance_miles'] = trips_th_with_pop['distance_miles']//distance_step*distance_step
trips_th_with_pop['duration_minutes'] = trips_th_with_pop['duration_minutes']//duration_step*duration_step

trips_th_with_pop = trips_th_with_pop[['person_id','travel_purpose','target_amenity','mode',
'distance_miles','duration_minutes','age','individual_income','household_size','family_structure','vehicles',
'income_group','age_group']]

trips_th_with_pop.head()

for idx,row in trips_th_with_pop.iterrows():
    person_description = f"A {row['age']} year old person, living in a {row['family_structure']} family with {row['household_size']} members. The person has {row['vehicles']} vehicles and an annual income of {row['individual_income']} dollars."
    trips_th_with_pop.loc[idx,'description'] = person_description
    cypher_profile = f"A {row['age_group']} person, whose income is {row['income_group']}"
    trips_th_with_pop.loc[idx,'cypher_profile'] = cypher_profile

trips_th_with_pop.head()

# remove age_group == 'Teen', because there are not enough data points
trips_th_with_pop = trips_th_with_pop[trips_th_with_pop['age_group']!='Teen']


trips_th_with_pop.loc[0,'cypher_profile']
