==> notebook/06_experiment.ipynb <==
from tqdm import tqdm

import pandas as pd
import os
import sys
sys.path.append(os.path.dirname(os.path.abspath('.')))

desire = 'Eat'
test_data_path = f'../data/test/{desire}.csv'
test_df = pd.read_csv(test_data_path, index_col=False)
test_df.head(2)

# Witout Context


from model import ChoiceModel
choice_model = ChoiceModel(
    data_dir='../data', desire=desire, sample_num=0, skip_init=True)

for idx, row in tqdm(test_df.iterrows(), total=len(test_df)):
    try:
        profile = row[['person_id', 'age', 'individual_income',
                       'household_size', 'family_structure', 'vehicles', 'cypher']]
        profile = profile.to_dict()
        amenity_final_choice, mode_final_choice = choice_model.infer_without_context(
            profile=profile, mode='experiment')
    except Exception as e:
        print(e)
choice_model.save_logs()

# With Context


from model import ChoiceModel
num_sample = 1000
choice_model = ChoiceModel(
    data_dir='../data', desire=desire, sample_num=num_sample, skip_init=False)

for idx, row in tqdm(test_df.iterrows(), total=len(test_df)):
    try:
        profile = row[['person_id', 'age', 'individual_income',
                        'household_size', 'family_structure', 'vehicles', 'cypher']]
        profile = profile.to_dict()
        amenity_final_choice, mode_final_choice = choice_model.infer(
            profile=profile, mode='experiment')
    except Exception as e:
        pass
choice_model.save_logs()

# from model import ChoiceModel

# for i in range(4, 11):
#     num_sample = i*100
#     choice_model = ChoiceModel(
#         data_dir='../data', desire=desire, sample_num=num_sample, skip_init=False)

#     for idx, row in tqdm(test_df.iterrows(), total=len(test_df)):
#         try:
#             profile = row[['person_id', 'age', 'individual_income',
#                            'household_size', 'family_structure', 'vehicles', 'cypher']]
#             profile = profile.to_dict()
#             amenity_final_choice, mode_final_choice = choice_model.infer(
#                 profile=profile, mode='experiment')
#         except Exception as e:
#             pass
#     choice_model.save_logs()
